fixing rabbitmq probes
