--- Dockerfile.CPA
@@ -1,18 +1,0 @@
-FROM python:3.9-slim
-
-WORKDIR /app
-
-# Copy only the necessary files to the container
-COPY ./autoscaler /app/autoscaler
-
-# Install dependencies from the requirements file located in the autoscaler subdirectory
-RUN pip install --no-cache-dir -r ./autoscaler/requirements.txt
-
-# Expose the desired port
-EXPOSE 3000
-
-# Set the working directory to the autoscaler directory
-WORKDIR /app/autoscaler
-
-# Run the application
-CMD ["python", "./autoscaler-request.py"]
