+++ app/consumer.py
@@ -1,11 +1,14 @@
-from asyncio import sleep
-import pika
+import threading
+import time
+from flask import Flask, g, jsonify, request
 import json
 import io
 import base64
 import logging
 from PIL import Image
+from prometheus_client import REGISTRY, Counter, Histogram, Summary, generate_latest, start_http_server
 from postgresConnector import PostgresConnectionManager
+from rabbitmqConnector import RabbitMQConnectionManager
 from classifier import ImageClassifier
 from os import environ, path
 
@@ -17,6 +20,12 @@ db_user = environ.get('DB_USER', 'root')
 db_password = environ.get('DB_PASSWORD', 'password')
 rabbitmq_host = environ.get('RABBITMQ_HOST', 'localhost')
 rabbitmq_queue = environ.get('RABBITMQ_QUEUE', 'requests_queue')
+rabbitmq_username = environ.get('RABBITMQ_USERNAME', 'flask')   
+rabbitmq_password = environ.get('RABBITMQ_PASSWORD', 'flask')
+app_port = int(environ.get('PORT', 5000))
+metrics_port = int(environ.get('METRICS_PORT', 8000))
+
+app = Flask(__name__)
 
 # Initialize the database manager
 db_manager = PostgresConnectionManager(
@@ -32,61 +41,60 @@ model_path = path.join(path.dirname(__file__), '../models/resnet18.pth')
 label_path = path.join(path.dirname(__file__), '../data/imagenet_classes.txt')
 classifier = ImageClassifier(model_name='resnet18', model_path=model_path, label_path=label_path)
 
-# RabbitMQ connection setup
-class RabbitMQConnectionManager:
-    def __init__(self, host, queue_name, max_retries=5, retry_delay=2):
-        self.host = host
-        self.queue_name = queue_name
-        self.max_retries = max_retries
-        self.retry_delay = retry_delay
-        self.connection = None
-        self.channel = None
-
-    def connect(self):
-        retries = 0
-        while retries < self.max_retries:
-            try:
-                self.connection = pika.BlockingConnection(pika.ConnectionParameters(host=self.host))
-                self.channel = self.connection.channel()
-                self.channel.queue_declare(queue=self.queue_name, durable=True)
-                logging.info("Successfully connected to RabbitMQ")
-                return
-            except pika.exceptions.AMQPConnectionError as e:
-                retries += 1
-                logging.warning(f"Connection attempt {retries} failed: {e}")
-                sleep(self.retry_delay)
-            except Exception as e:
-                logging.error(f"Failed to connect to RabbitMQ: {e}")
-                raise e
-        logging.error("Exceeded maximum retries, could not connect to RabbitMQ")
-        raise ConnectionError("Could not connect to RabbitMQ after multiple attempts.")
-
-    def get_channel(self):
-        if self.connection is None or self.connection.is_closed or self.channel is None or self.channel.is_closed:
-            logging.info("No active RabbitMQ connection. Attempting to connect/reconnect.")
-            self.connect()
-        return self.channel
-
-    def close(self):
-        if self.connection and not self.connection.is_closed:
-            try:
-                self.connection.close()
-                logging.info("RabbitMQ connection closed.")
-            except Exception as e:
-                logging.error(f"Error closing RabbitMQ connection: {e}")
-                raise e
-
-    def __enter__(self):
-        self.connect()
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.close()
-
+# initialize the RabbitMQ connection manager
 rabbitmq_manager = RabbitMQConnectionManager(
     host=rabbitmq_host,
-    queue_name=rabbitmq_queue
+    queue_name=rabbitmq_queue,
+    rabbitmq_username=rabbitmq_username,
+    rabbitmq_password=rabbitmq_password
 )
+
+route_hit_counter = Counter('route_hits', 'Count of hits to routes', ['route'])
+REQUEST_TIME = Summary('request_processing_seconds', 'Time spent processing request', registry=REGISTRY)
+REQUEST_COUNT = Counter('request_count', 'Total number of requests', ['method', 'endpoint', 'http_status'], registry=REGISTRY)
+REQUEST_LATENCY = Histogram('request_latency_seconds', 'Request latency', ['method', 'endpoint'], registry=REGISTRY)
+
+@app.before_request
+def before_request():
+    g.start_time = time.time()
+
+@app.after_request
+def after_request(response):
+    latency = time.time() - g.start_time
+    REQUEST_LATENCY.labels(request.method, request.path).observe(latency)
+    REQUEST_COUNT.labels(request.method, request.path, response.status_code).inc()
+    return response
+
+@app.errorhandler(Exception)
+def handle_exception(e):
+    response = {
+        "error": str(e),
+        "message": "An unexpected error occurred. Please try again later."
+    }
+    return jsonify(response), 500
+
+
+@app.route('/')
+@REQUEST_TIME.time()
+def index():
+    route_hit_counter.labels(route='/').inc()
+    return "<h1>Hello, worker here!</h1>"
+
+@app.route('/health')
+def health():
+    route_hit_counter.labels(route='/health').inc()
+    return "OK", 200
+
+@app.route('/metrics')
+def metrics():
+    route_hit_counter.labels(route='/metrics').inc()
+    return generate_latest(REGISTRY), 200, {'Content-Type': 'text/plain; charset=utf-8'}
+
+def start_flask_app():
+    app.run(host='0.0.0.0', port=app_port)
+
+def start_metrics_server():
+    start_http_server(metrics_port)
 
 def callback(ch, method, properties, body):
     try:
@@ -126,6 +134,16 @@ def start_consuming():
     logging.info("Starting to consume messages...")
     channel.start_consuming()
 
+def main():
+    # Start the Prometheus metrics server in a separate thread
+    threading.Thread(target=start_metrics_server).start()
+
+    # Start the Flask app in a separate thread
+    threading.Thread(target=start_flask_app).start()
+
+    # Start RabbitMQ consumer in the main thread
+    start_consuming()
+
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO)
-    start_consuming()
+    main()
