+++ deployments/apps/app.yaml
@@ -18,7 +18,7 @@ spec:
     spec:
       containers:
       - name: flask-app
-        image: gunicorn-app:latest
+        image: flask-app:latest
         imagePullPolicy: IfNotPresent
         ports:
         - containerPort: 5000
@@ -40,6 +40,24 @@ spec:
           value: "5000"
         - name: METRICS_PORT
           value: "8000"
+        - name: DB_NAME
+          value: "resnet18_db"
+        - name: DB_USER
+          value: "root"
+        - name: DB_PASSWORD
+          value: "password"
+        - name: DB_HOST
+          value: "postgres-service"
+        - name: DB_PORT
+          value: "5432"
+        - name: RABBITMQ_HOST        
+          value: "rabbitmq"
+        - name: RABBITMQ_QUEUE
+          value: "requests_queue"
+        - name: RABBITMQ_USERNAME
+          value: "flask" 
+        - name: RABBITMQ_PASSWORD
+          value: "flask"  
 ---
 apiVersion: v1
 kind: Service
