+++ deployments/rabbitmq/rabbitmq.yaml
@@ -1,3 +1,27 @@
+apiVersion: v1
+kind: PersistentVolume
+metadata:
+  name: rabbitmq-pv
+spec:
+  capacity:
+    storage: 1Gi  
+  accessModes:
+    - ReadWriteOnce  
+  persistentVolumeReclaimPolicy: Retain
+  hostPath:
+    path: "/mnt/data/rabbitmq"  
+---
+apiVersion: v1
+kind: PersistentVolumeClaim
+metadata:
+  name: rabbitmq-pvc
+spec:
+  accessModes:
+    - ReadWriteOnce
+  resources:
+    requests:
+      storage: 1Gi  
+---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -9,7 +33,7 @@ spec:
   selector:
     matchLabels:
       app: rabbitmq
-  template:
+  template: 
     metadata:
       labels:
         app: rabbitmq
@@ -22,62 +46,48 @@ spec:
         - containerPort: 15672  # Management UI Port
         volumeMounts:
         - name: rabbitmq-storage
-          mountPath: /var/lib/rabbitmq
+          mountPath: /var/lib/rabbitmq  
+        env:
+        - name: RABBITMQ_DEFAULT_USER
+          value: "admin"  
+        - name: RABBITMQ_DEFAULT_PASS
+          value: "admin"  
+        - name: RABBITMQ_USERS
+          value: "flask:flask:administrator"  
+
+        # Liveness and Readiness Probes
         livenessProbe:
-          httpGet:
-            path: /api/healthchecks/node
-            port: 15672
-          initialDelaySeconds: 60
-          periodSeconds: 30
+          tcpSocket:
+            port: 5672
+          initialDelaySeconds: 20
+          timeoutSeconds: 5
+
         readinessProbe:
-          httpGet:
-            path: /api/healthchecks/node
-            port: 15672
-          initialDelaySeconds: 30
-          periodSeconds: 15
+          tcpSocket:
+            port: 5672
+          initialDelaySeconds: 10
+          timeoutSeconds: 5
+        
+
       volumes:
       - name: rabbitmq-storage
         persistentVolumeClaim:
           claimName: rabbitmq-pvc
-
 ---
 apiVersion: v1
 kind: Service
 metadata:
   name: rabbitmq
+  labels:
+    app: rabbitmq
 spec:
-  ports:
-  - port: 5672
-    targetPort: 5672
-    name: amqp
-  - port: 15672
-    targetPort: 15672
-    name: management
   selector:
     app: rabbitmq
-  type: ClusterIP   # Changed to ClusterIP for internal communication
-
----
-apiVersion: v1
-kind: PersistentVolume
-metadata:
-  name: rabbitmq-pv
-spec:
-  capacity:
-    storage: 1Gi
-  accessModes:
-    - ReadWriteOnce
-  hostPath:  # Note: This is specific to a node. Consider using dynamic provisioning in production.
-    path: /mnt/data/rabbitmq
-
----
-apiVersion: v1
-kind: PersistentVolumeClaim
-metadata:
-  name: rabbitmq-pvc
-spec:
-  accessModes:
-    - ReadWriteOnce
-  resources:
-    requests:
-      storage: 1Gi
+  ports:
+  - name: amqp
+    port: 5672  
+    targetPort: 5672
+  - name: management
+    port: 15672  
+    targetPort: 15672
+  type: NodePort
