--- docker-compose.yml
@@ -1,13 +1,0 @@
-version: '3.8'
-services:
-  flask-app:
-    container_name: flask-app
-    image: resnet18-flask-app:latest
-    ports:
-     - "${PORT}:${PORT}"
-     - "${METRICS_PORT}:${METRICS_PORT}"
-    volumes:
-     - .:/code
-    environment:
-      - PORT=${PORT}
-      - METRICS_PORT=${METRICS_PORT}
