+++ docker/consumer/requirements.txt
@@ -1,5 +1,12 @@
-pika==1.2.0
-Pillow==9.5.0
-psycopg2-binary==2.9.5
-asyncio==3.4.3
-requests==2.31.0
+prometheus_client==0.20.0
+torch==2.3.1
+torchvision==0.18.1
+Pillow==10.3.0
+requests==2.32.3
+aiohttp==3.9.5
+gunicorn==22.0.0
+Flask==3.0.3
+psycopg2-binary==2.9.9
+pika==1.3.2
+Flask-SQLAlchemy
+Flask-Migrate
