--- docker/producer/Dockerfile.Worker
@@ -1,15 +1,0 @@
-FROM python:3.8-slim
-
-ENV PORT 5000
-ENV METRICS_PORT 8000
-
-WORKDIR /app
-
-COPY ./docker/producer/requirements.txt requirements.txt
-RUN pip install -r requirements.txt
-
-COPY . .
-
-EXPOSE 5000
-
-CMD ["python", "worker_gunicorn.py"]
